fix: invoke plugin.uninstall in startup task
