+++ GalgameManager/Contracts/Services/IPluginService.cs
@@ -6,24 +6,6 @@ namespace GalgameManager.Contracts.Services;
 public interface IPluginService
 {
     /// <summary>
-    /// 是否有正在等待取消加载/删除的插件
-    /// </summary>
-    public bool PluginOffloadInProgress { get; }
-    
-    /// <summary>
-    /// 立即设置插件的关联数据
-    /// </summary>
-    /// <param name="plugin"></param>
-    /// <param name="data"></param>
-    public void PluginSetData(PluginX plugin, string? data); 
-    
-    /// <summary>
-    /// 立即删除插件的关联数据
-    /// </summary>
-    /// <param name="plugin"></param>
-    public void PluginDeleteData(PluginX plugin); 
-    
-    /// <summary>
     /// 新增插件，注意捕获异常
     /// </summary>
     /// <param name="path"></param>
@@ -42,9 +24,9 @@ public interface IPluginService
     public Task DeletePluginAsync(PluginX plugin, bool deleteData);
     
     public Task<ObservableCollection<PluginX>> GetAllPluginsAsync();
-    
+
     public Task InitAsync();
-    
+
     /// <summary>
     /// 加载某个插件，如果已经加载则不操作，注意捕获异常
     /// </summary>
@@ -52,11 +34,29 @@ public interface IPluginService
     /// <param name="load">是否要加载插件，若设置为false则只把插件加到插件列表里而不加载（初始化插件表时用）</param>
     /// <returns></returns>
     public Task LoadPluginAsync(PluginX plugin, bool load);
+    
+    /// <summary>
+    /// 立即删除插件的关联数据
+    /// </summary>
+    /// <param name="plugin"></param>
+    public void PluginDeleteData(PluginX plugin); 
 
     /// <summary>
     /// 插件存放目录
     /// </summary>
     public DirectoryInfo PluginDir { get; }
+    
+    /// <summary>
+    /// 是否有正在等待取消加载/删除的插件
+    /// </summary>
+    public bool PluginOffloadInProgress { get; }
+    
+    /// <summary>
+    /// 立即设置插件的关联数据
+    /// </summary>
+    /// <param name="plugin"></param>
+    /// <param name="data"></param>
+    public void PluginSetData(PluginX plugin, string? data); 
     
     /// <summary>
     /// 当调用插件功能抛出异常时可以调用此方法提醒用户（内部会调用infoService发送一个Event）
