+++ GalgameManager/Models/BgTasks/LoadPluginTask.cs
@@ -28,7 +28,7 @@ public class LoadPluginTask : BgTaskBase
             PluginX plugin = plugins[i];
             if (plugin.ToDelete)
             {
-                DeletePlugin(plugin);
+                await DeletePlugin(plugin);
                 continue;
             }
             
@@ -47,7 +47,7 @@ public class LoadPluginTask : BgTaskBase
         ChangeProgress(1, 1, string.Empty, false);
         return;
 
-        void DeletePlugin(PluginX plugin)
+        async Task DeletePlugin(PluginX plugin)
         {
             try
             {
@@ -63,6 +63,10 @@ public class LoadPluginTask : BgTaskBase
                     if (Directory.Exists(plugin.Path)) 
                         Directory.Delete(plugin.Path, true);
                 }
+
+                // 如果出现文件占用问题，则不删除 plugin，等待下次重启再删除。
+                // 如果 plugin 自身的 OnUninstall 超时或者报错，则正常删除这个插件。
+                try { await plugin.ExecuteUninstallWithTimeoutAsync(); } catch(Exception) { /* Ignore */ }
                 if (plugin.ToDeleteData) _pluginService.PluginDeleteData(plugin);
                 db.Delete(plugin.Id);
             }
